Major changes to CSS handling in Iodine: 1. CSS files are now in the styles directory 2. Stylesheets are split up into seperate files that are 'compiled' by the css module 3. User styles (modern, ocean, etc.) extend the default style. (There is no 'global' stylesheet, you override what you want) 4. User styles can either 'extend' or 'replace' selectors by the @extend and @replace (default is replace): 	Replacing completely redefines the attributes for a selector 	Extending uses as many attributes from the parent style as possible 5. User styles can be based on other user styles, which eventually lead back to 'default' 	The autumn style is based on the ocean style 6. The CSS module will only 'recompile' the css files when necessary (based on directory/file modification time) 7. Styles are no longer specified in config.ini.  The CSS module has a static method to list styles.
I'll try and add more detailed documentation to the css module and relevant classes

This should make it much easier to create non style-specific css files (eg. positioning of elements in a module) and have them affect all styles.

The styles that are not yet implemented are still in the www/styles directory.

I left the login style alone, since it would be complicated to use the css module to serve a stylesheet to a user that is not authenticated.

I probably forgot a lot of stuff, but here goes!