Make special activities comprehensible in the 8th box, and begin antistalker support.
Please note that this permission bit is different from all the others:
First, it must be implemented by the Iodine application because we do not have
8th-period data in LDAP.  We should discuss the possibility of making such a
move, weigh the potential performance loss against the security and
interoperability gains, and make a decision at some future point.  For now,
we'll hack in applications-layer permissions.
Second, this bit is backwards.  It is a 'hide' bit, not a 'show' bit.  This is
so that we don't have to add another piece of info to every user in the
database, teachers will work correctly if they become activity members, etc.
Just bear with it.