Added expiration date for news *MYSQL UPDATE*
In this update, everything needed for the expiration date of news is included. Currently, the interface for date selection for adding a news item and editing one are slightly different (the edit uses the MySQL datetime syntax, add uses drop-downs); this will hopefully change once I learn more smarty (Est. time: 1-3 weeks).

MySQL CHANGE:
ALTER TABLE news ADD COLUMN expire DATETIME NULL DEFAULT NULL;

(This should work: it's goal is to add in a column called expire that represents a datetime where it can be null and the default is null).

Forthcoming changes:
* Move javascript in templates/utils/calendar.tpl (req. JavaScript voodoo magic)
* Make better calendar.tpl (add time option; default value)