Minor changes: fixed javascript bug and added base tag.
The javascript bug was a mere annoyance that the code produced an error due to the inaccessibility of event if the mouse was moved before the page finished loading.

The base tag means that "[<$I2_ROOT>]" does not need to be prepended onto every link; it has been tested and shown to work in elinks, IE, and Firefox (all somewhat briefly). The main intent of the addition of the base tag is to aid in scripting so that pure scripts, e.g. scratchpad, do not need to have script tags to pass the I2_ROOT page into the .js file. In short, its use with respect to normal links is weakly encouraged (for consistency's sake), but it facilitates ease of use in scripting.