Mammoth Polls Update 1.8 (Group Permissions fixes)
This poll update obviates the groups permissions, and makes the group access clean.
I still dislike the lack of an automagical "remove selected groups from other
dropdowns", but that would be too complex. After all, we do want polls to be
complete sometime in the near future!

Non-JS group automagic is a little broken, but it is a step above what used to
exist: groups can be accessed without JS now, even if they have to be added
one at a time and cannot be set on polls_add.tpl