Adding 8th office feature: colored block labels.
So I violate a few manuals of good web page design (but I2 forces me to), the feature is probably going to horribly crash and burn in IE, and I use a feature not even supported in FF trunk.

Also, <insert diatribe on the crappiness of EighthPrint here>. At least brj didn't write it, adeason did.