A few changes here, (I know that's bad form, but still):
1. Add a new type marker for mysql queries, use %X for escaped but unquoted
strings in queries. Only use this if absolutely needed, e.g. you're using a
term in a regex, and it would break the regex if the quotes were there.
2. Add header stuff so that teachers (or students!) who are hosting 8th period
activities have it added to their header with links, which still have to be
changed to go to an attendance taking page.
3. Make num_members() in EighthActivity _actually work_.
4. In EighthSponsor change get_schedule() so that it's more efficient and has
less work in the php.
5. In EighthSponsor add get_schedule_on() that gives the schedule of a given
sponsor or array of sponsors on a given date.
6. In eighth, make vp_sponsor() use a switch statement and redirect bad
requests back to the main page of vp_sponsor.
7. Add a userid variable to the eighth_sponsors sql table definition.
8. Minimal additions for graduate headers.
9. In vp_sponsor.tpl, use the newly-functional num_members() instead of a set
of loops and a bunch of object initializations in the smarty.