Fix API signup_activity (hopefully)
Right now it appears that through the API, you can sign up for cancelled or even full activities (eek). This is probably a bad thing. This is my best effort to fix the problem based on looking at other code, although I have little knowledge of PHP and no way to test it!